+++ components/UI/loading/LoadingSpinner.tsx
@@ -1,6 +1,6 @@
 import { FadeLoader } from "react-spinners";
 
-const LoadingSpinner = ({ text }: { text: string }) => {
+const LoadingSpinner = ({ text }: { text?: string }) => {
   return (
     <div className="flex flex-col items-center justify-center h-full">
       <p className="text-2xl mb-4">{text}</p>
